python/requirements.txt
beautifulsoup4==4.12.3
pymongo==4.8.0
python-dotenv==1.0.1
Requests==2.32.3
tqdm==4.66.4
pandas==2.3.1